+++ components/content/index.vue
@@ -608,7 +608,7 @@ export default {
 
 .listswiper /deep/ .mescroll-totop-all {
   box-sizing: border-box;
-  z-index: 2901;
+  z-index: 999;
   position: fixed;
   bottom: 65px;
   width: 50px;