+++ CHANGELOG.md
@@ -4,6 +4,15 @@ toc: hidden
 ---
 
 
+### 0.0.36-未发版
+
+`2020-08-26`
+
+- Feat
+
+- Fix
+  - 1、修复`content`组件滑动内容区高度兼容性问题，通过`vw`跟手机实际宽度`px`换算真实的`px`(`0.0.34`版本`Fix-1`的继续修复)。
+
 ### 0.0.35
 
 `2020-08-07`
@@ -17,7 +26,7 @@ toc: hidden
 
 - Fix
   - 1、修复`drop-select-item`组件`pickerTitle`属性失效问题。
-  - 2、`al-content`组件返回顶部按钮`mescroll-totop-all`样式`z-index`调整为`999`,解决返回顶部按钮遮挡弹出窗等`popup`弹出层
+  - 2、`content`组件返回顶部按钮`mescroll-totop-all`样式`z-index`调整为`999`,解决返回顶部按钮遮挡弹出窗等`popup`弹出层
   - 3、新增`dataListId`表数据标签`id`,修复当存在两个列表页面跳转时，因为默认的id相同，可能会导致，一个页面的接口请求因为延迟，当跳转到另一个页面，数据错误渲染在另一个页面，此时就需要设置两个页面不同的`dataListId`参数，防止此问题发生
   - 4、dom.js `allScreenHeight`属性值由视窗高度更改为网页高度
   - 5、dom.js `dealPxToVw`属性增增加`autoUnit`值，默认为`false`不拼接单位。当给定`autoUnit`-`false`并且视窗宽度大于1500时候（即类似pc端），给定`px`单位，不做vw的转换